fix: avoid unnecessarily chunking upstream requests
Although we don't really know why, it seems that making GET requests with
transfer-encoding: chunked (which happens automatically if an iterable is
passed as the request body to urllib3 and there is no content-length) causes
some origin servers to drop the GET requests before their responses have been
fully returned to the client.

This change addresses this by only using an iterable as the request body if we
need to. It also uses the 'iter' function to save a few lines of code while
we're at it.